Updated moodAnalyzer to check for errors.
If the text analyzed wasn't sufficient or detailed enough, it will return an empty list which would then be interpreted by the GUI and send an error message to the user, asking them to make a more descriptive journal entry. This was added because sometimes the text input would return no potential emotions if it was short.