Updated PlaylistGenerator.py
If there are less than 250 liked songs, we get the users top artists and return their top tracks instead. If that is less than 50 songs then we return the users top 50 tracks. If no songs pass the emotion_tracks, then we return "NO SONGS" which leads to a certain UI.